api_server/requirements.txt
celery==5.2.3
Django==3.2.11
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
